Consolidate Parquet Metadata handling
