+++ datafusion/datasource-parquet/src/mod.rs
@@ -21,6 +21,7 @@
 
 pub mod access_plan;
 pub mod file_format;
+pub mod metadata;
 mod metrics;
 mod opener;
 mod page_filter;
